Symbol.Message now Symbol.Signal / Event.owner
The Signal symbol identifies whether an "argument object" represents a
Message, Event, Broadcast, etc.

We should use the Signal symbol to tie the propagation, if any, to the
argument.

(2) Instead of putting the selection on the event (in some cases!), all
events have the owner property and owner.selection is available.