IMPORTANT CHANGE - Signals
Receiver: just go to its controller.
Controller: Controller calls execute to execute for each function called
on a receiver.  Controller.execute() is a method Call.
Processor: Extends Controller to execute its actions [to be called
"instructions"?]

Sender: A Sender is a *Controller*.  It is called after$process.

So all smarts in signals are within the Controllers.