Engine.extend: now supports loading & compilation of declaration values.
There are warnings about Loading cycles that need to be reviewed.

Due to this change, it doesn't make sense for modules to define the
packages (because they would get loaded when creating the module & again
when compiling. [actually it can be overcome by not loading when
compiling module packages, anyway.].  Regardless, packages are now
defined in the conf object and the module is as well.  The next commit
contains module.mjs removal.  Need to update documentation.