Important: Public pattern
You can now define packages with a $public property. This enables a
package scope.  When Module compiles a package with a $public
property, the public property will be compiled instead.  One of the
biggest things to leverage here is the abilty to manipulate functions
since the entire package scope is exported.