Fixed super...
super was failing when the chain extended beyond one super.  The reason was we were relying on this[name] to get the method.  "this" restarts the chain.  For super to work right we need a reference to the method itself.  so when calling super, the method must be named in the declared so it can be referenced, e.g.
{
  method: function method(x) {
    this.super(method, x); // pass in the current method.
  }
}