ops clarification & renaming...
An instruction is an assembler instruction.  It can assemble to a machine operation.  For example, "add" can assemble to one of two ops based on the arguments (e.g. reg vs. immed) in the assembly instruction.

So, "op" is generally renamed to instruction.

Added Code & Data instructions as type = "directive".