** pls read ** concrete editor subtypes created (modellers, viewers, editors are no longer used).
a number of imports were missing the ".js" extension (somehow). The editor.ts was created via a ts refactoring and that accounts for most of it.  However, there were others missing the .js ext and the TS compiler didn't complain, probably because those modules only defined interfaces and therefore no .js to generate so it didn't care.