BIG SEMANTIC CHANGE:
NodeContent is now the "View".  Instead of having NodeContent.node we have View.view which returns a generic T for the view.

Also, Content.contents<T> is now contents<Text> so that the view is not coupled to the contents - in practice this means we can now declare View<ELE> rather than View<NODE> in the implementations.